requirements.txt
cffi==1.17.1
cryptography==43.0.3
pillow==10.4.0
pycparser==2.22
pyotp==2.9.0
pypng==0.20220715.0
qrcode==7.4.2
typing-extensions==4.12.2